major structure changes, added some basic tests
