+++ advice.js
@@ -1,0 +1,6 @@
+function advice(fn){
+  fn.advice = 1;
+  return fn;
+}
+
+module.exports = advice;
