+++ factory.js
@@ -1,0 +1,7 @@
+function factory(_type){
+  return function () {
+    return new _type;
+  }
+}
+
+module.exports = factory;
