+++ package.json
@@ -2,7 +2,7 @@
   "name": "kaop",
   "version": "3.0.0",
   "description": "simple js OOP lib helper with IoC extensions",
-  "main": "type.js",
+  "main": "index.js",
   "scripts": {
     "test": "jest"
   },
@@ -21,5 +21,8 @@
   "bugs": {
     "url": "https://github.com/k1r0s/kaop/issues"
   },
-  "license": "MIT"
+  "license": "MIT",
+  "devDependencies": {
+    "jest": "^21.2.1"
+  }
 }
